changed: changed code in some places to remove warnings generated by Visual Studio added:   file win64_assemble.bat to compile (assemble) *.asm file for Win64
git-svn-id: svn://ttmath.org/publicrep/ttmath/trunk@1045 e52654a7-88a9-db11-a3e9-0013d4bc506e

+++ ttmath/ttmathuint.h
@@ -3319,7 +3319,7 @@ public:
 		if( negative )
 			result = '-';
 
-		digits_d  = table_id; // for not making an overflow in uint type
+		digits_d  = static_cast<double>(table_id); // for not making an overflow in uint type
 		digits_d *= TTMATH_BITS_PER_UINT;
 		digits_d += index + 1;
 		digits_d *= ToStringLog2(b);
